// apps/aagora-core/types/types.ts
export interface UnitOfMeasure {
    id: string;
    name: string;
    abbreviation: string;
    magnitude: string;
}